+++ src/main/java/racingcar/controller/RandomRaceController.java
@@ -32,5 +32,4 @@ public class RandomRaceController {
             }
         }
     }
-
 }
