+++ src/main/java/racingcar/validator/TryNumbersValidator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.validator;
 
 public class TryNumbersValidator {
