+++ src/test/java/racingcar/CarNameValidatorTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.validator.CarNameValidator;
 
 class CarNameValidatorTest {
     private CarNameValidator carNameValidator;
